Welcome popup developed.

+++ behaviors/MerchiumClientBehavior.php
@@ -4,7 +4,9 @@ namespace app\behaviors;
 
 use Yii;
 use yii\base\Behavior;
+use yii\helpers\Html;
 use yii\helpers\Url;
+use yii\helpers\Json;
 use yii\db\ActiveRecord;
 use app\models\StoreOption;
 
@@ -43,6 +45,8 @@ class MerchiumClientBehavior extends Behavior
 
         $options = $store->getStoreOptionsByName();
 
+        $scripts = [];
+
         if (isset($options['snowfall_enable'])) {
 
             if (!empty($options['snowfall_script_tag_hash'])) {
@@ -70,23 +74,18 @@ class MerchiumClientBehavior extends Behavior
                 }
 
                 // Template hook
-                $client->createRequest('template_hooks', [
-                    'hookname' => 'index:scripts',
-                    'type' => 'post',
-                    'body' => '
-                        {literal}
-                        <script type="text/javascript">
-                        (function(_, $) {
-                            $(function(){
-                                setTimeout(function(){
-                                    $(document).snowfall({flakeCount: 100, maxSpeed: 5, maxSize: 4, shadow: true});
-                                }, 500);
-                            });
-                        }(Tygh, Tygh.$));
-                        </script>
-                        {/literal}
-                    ',
-                ]);
+                $scripts[] = '
+                    var snowfallLoad = function(){
+                        if ($.snowfall) {
+                            $(document).snowfall({flakeCount: 100, maxSpeed: 5, maxSize: 4, shadow: true});
+                        } else {
+                            setTimeout(function(){
+                                snowfallLoad();
+                            }, 300);
+                        }
+                    };
+                    snowfallLoad();
+                ';
 
             // Disable
             } else {
@@ -95,10 +94,51 @@ class MerchiumClientBehavior extends Behavior
                 $client->deleteRequest('script_tags/' . $hash_option->value);
                 $hash_option->delete();
 
-                // Template hook
-                $client->deleteRequest('template_hooks/index:scripts');
-
             }
+        }
+
+        if (isset($options['welcome_popup_enable'])) {
+
+            if ($options['welcome_popup_enable']->value) {
+                $welcome_popup_data = [
+                    'title' => Html::encode($options['welcome_popup_title']->value),
+                    'content' => Html::encode($options['welcome_popup_content']->value),
+                ];
+                $welcome_popup_data_json = Json::encode($welcome_popup_data);
+
+                $scripts[] = "
+                    $(function(){
+                        var welcome_popup_data = {$welcome_popup_data_json};
+                        if (!$.cookie.get('welcome_popup_displayed')) {
+                            $.cookie.set('welcome_popup_displayed', true);
+                            $.ceNotification('show', {
+                                type: 'I',
+                                title: welcome_popup_data.title,
+                                message: '<div style=\"margin: 0 20px 0 20px;\">' + welcome_popup_data.content + '</div>',
+                                message_state: 'S'
+                            }, 'popupmodal');
+                        }
+                    });
+                ";
+            }
+
+        }
+
+        if ($scripts) {
+            $client->createRequest('template_hooks', [
+                'hookname' => 'index:scripts',
+                'type' => 'post',
+                'body' => '
+                    {literal}
+                    <script type="text/javascript">
+                    (function(_, $) {
+                    ' . implode(PHP_EOL, $scripts) . '
+                    }(Tygh, Tygh.$));
+                    </script>
+                    {/literal}',
+            ]);
+        } else {
+            $client->deleteRequest('template_hooks/index:scripts');
         }
 
     }
